% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dinkageverden.R
\name{dinkageverden}
\alias{dinkageverden}
\title{Hent prisen p책 en vare p책 dinkageverden.dk}
\usage{
dinkageverden(link)
}
\arguments{
\item{link}{link}
}
\description{
Hent prisen p책 en vare p책 dinkageverden.dk
}
